rotary encoders working (ISR's implemented)

+++ src/OptionsTab.h
@@ -11,16 +11,12 @@ class OptionsTab : public Tab
         uint8_t verticalIndex;
         uint8_t currentCLK;
         uint8_t lastCLK;
-        uint8_t horozontalIndex;
-        uint8_t prevHorozontalIndex;
+        int32_t horozontalIndex;
+        int32_t prevHorozontalIndex;
         uint64_t timer;
         uint8_t colourIndex;
         uint8_t altitudeOffset;
         uint8_t volume;
-        //uint8_t index1;
-        //uint8_t index2;
-        //uint8_t select;
-        //bool selected; 
         String dataToBeSaved;
     public:
        	OptionsTab(Adafruit_TFTLCD *ptft, ConfigData *pPipData);
